[#7] The return type of MinIDSearch::result() should be int instead of double

See merge request astuff/autoware.ai/common!27

+++ libwaypoint_follower/src/libwaypoint_follower.cpp
@@ -229,7 +229,7 @@ public:
       val_min_ = v;
     }
   }
-  const double result() const
+  const int result() const
   {
     return idx_min_;
   }
